+++ app/javascript/components/game/map/MapDisplay.tsx
@@ -583,6 +583,7 @@ export default function MapDisplay({
       setMoveTrack(<MoveTrackOverlay map={map} scale={scale} mapScale={mapScale ?? 1} updateCallback={() =>
                                        { counterCallback(); updateCallback() }}
                                      svgRef={svgRef as React.MutableRefObject<HTMLElement>} />)
+      counterCallback()
     } else {
       // TODO: if last sig action was fire, check for displace
       setMoveTrack(undefined)